+++ mockify/core/_call.py
@@ -119,7 +119,7 @@ class Call:
 
     @staticmethod
     def _format_params(*args, **kwargs):
-        return _utils.format_args_kwargs(args, kwargs)
+        return _utils.ArgsKwargsFormatter().format(*args, **kwargs)
 
     @property
     def name(self):